+++ linux/litertlm_binaries.cmake
@@ -1,4 +1,4 @@
 set(LITERTLM_X86_64_URL
-    "https://files.pythonhosted.org/packages/e3/e3/eb08521ef440ca7967dc8cb057fc5701bc98adcd5cfb78098da8360ade41/litert_lm_api-0.15.0-py3-none-manylinux_2_27_x86_64.whl")
+    "https://files.pythonhosted.org/packages/1a/b5/52cdeda2ed557ce96dd1c627c25177eca3e31fe3219ae7f4ec5f28a1d777/litert_lm_api-0.17.0-py3-none-manylinux_2_27_x86_64.whl")
 set(LITERTLM_ARM64_URL
-    "https://files.pythonhosted.org/packages/60/6d/d7dbeee183fb96be60dc965e0675c4f54e68ae1662da442189332729be58/litert_lm_api-0.15.0-py3-none-manylinux_2_27_aarch64.whl")
+    "https://files.pythonhosted.org/packages/b6/63/f5d575f2c88f44a039c19ff915ec25c7ec4b1b46e5ee28407d4c1392a50f/litert_lm_api-0.17.0-py3-none-manylinux_2_27_aarch64.whl")