+++ windows/litertlm_binaries.cmake
@@ -1,2 +1,2 @@
 set(LITERTLM_X86_64_URL
-    "https://files.pythonhosted.org/packages/c8/41/d70672921bd97c50679845b825dd02bfec875faece2e1529102502b7e6ba/litert_lm_api-0.15.0-py3-none-win_amd64.whl")
+    "https://files.pythonhosted.org/packages/46/09/fab014fe4394a3f1d21e30622fbd4add89b69b0688382bc93aeedde007f2/litert_lm_api-0.17.0-py3-none-win_amd64.whl")